Refactor KafkaStreamSink.Provider to KafkaStreaMessaging
