+++ eagle-core/eagle-app/eagle-app-base/src/test/resources/application.conf
@@ -43,8 +43,8 @@
 		"store": "org.apache.eagle.metadata.service.memory.MemoryMetadataStore"
 	},
 	"application":{
-		"sink":{
-			"type": "org.apache.eagle.app.messaging.KafkaStreamSink"
+		"messaging":{
+			"type": "org.apache.eagle.app.messaging.KafkaStreamMessaging"
 			"boostrap.server":"localhost:9092"
 		}
 		"provider":{
