+++ eagle-server/src/main/resources/application.conf
@@ -85,7 +85,7 @@ metadata {
 # Eagle Application Configuration
 # ---------------------------------------------
 application {
-  sink {
+  messaging {
     type = org.apache.eagle.app.messaging.KafkaStreamSink
   }
   storm {
